Moved setupNewAccount scripts into its own folder
