+++ scripts/src/setupNewAccount/types.ts
